+++ CopilotKit/packages/react-core/src/hooks/use-copilot-chat.ts
@@ -1,11 +1,13 @@
 import { useMemo, useContext } from "react";
 import { CopilotContext } from "../context/copilot-context";
 import { Message, Function } from "@copilotkit/shared";
+import { SystemMessageFunction } from "../types";
 import { UseChatOptions, useChat } from "./use-chat";
 import { defaultCopilotContextCategories } from "../components";
 
 export interface UseCopilotChatOptions extends UseChatOptions {
-  makeSystemMessage?: (contextString: string) => string;
+  makeSystemMessage?: SystemMessageFunction;
+  additionalInstructions?: string;
 }
 
 export interface UseCopilotChatReturn {
@@ -20,6 +22,7 @@ export interface UseCopilotChatReturn {
 
 export function useCopilotChat({
   makeSystemMessage,
+  additionalInstructions,
   ...options
 }: UseCopilotChatOptions): UseCopilotChatReturn {
   const {
@@ -35,7 +38,7 @@ export function useCopilotChat({
 
     return {
       id: "system",
-      content: systemMessageMaker(contextString),
+      content: systemMessageMaker(contextString, additionalInstructions),
       role: "system",
     };
   }, [getContextString, makeSystemMessage]);
@@ -72,8 +75,12 @@ export function useCopilotChat({
   };
 }
 
-export function defaultSystemMessage(contextString: string): string {
-  return `
+export function defaultSystemMessage(
+  contextString: string,
+  additionalInstructions?: string,
+): string {
+  return (
+    `
 Please act as an efficient, competent, conscientious, and industrious professional assistant.
 
 Help the user achieve their goals, and you do so in a way that is as efficient as possible, without unnecessary fluff, but also without sacrificing professionalism.
@@ -91,5 +98,6 @@ Please assist them as best you can.
 You can ask them for clarifying questions if needed, but don't be annoying about it. If you can reasonably 'fill in the blanks' yourself, do so.
 
 If you would like to call a function, call it without saying anything else.
-`;
+` + (additionalInstructions ? `\n\n${additionalInstructions}` : "")
+  );
 }