+++ CopilotKit/packages/react-core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # ui
 
+## 0.18.0-alpha.0
+
+### Minor Changes
+
+- New Copilot ui
+
 ## 0.17.1
 
 ### Patch Changes